feat: add layout for site slug pages to avoid rerender on nav
