+++ apps/web/app/(dashboard)/@sidebar/sites/[...slug]/page.tsx
@@ -1,16 +1,3 @@
-import { getSiteAndTeam } from "@repo/database";
-import { AppSidebar } from "@/components/app-sidebar";
-
-type Params = Promise<{ slug?: string[] }>;
-
-export default async function Page({ params }: { params: Params }) {
-    const { slug } = await params;
-    let site;
-    if (slug && slug.length > 1) {
-        console.log("slugA", slug);
-        site = await getSiteAndTeam(slug[1]!);
-    }  else {
-        console.log("slugB", slug);
-    }
-    return <AppSidebar currentSite={site} currentTeam={site?.team ? site.team : undefined}/>
+export default function Page() {
+    return null;
 }
